# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tenants": {
      "post": {
        "operationId": "TenantsController_create",
        "summary": "Create a new church/tenant (super admin only)",
        "tags": ["Tenants"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTenantDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Tenant created. Call /auth/refresh for updated JWT." },
          "403": { "description": "Super admin access required" }
        }
      }
    },
    "/api/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "summary": "Get tenant details",
        "tags": ["Tenants"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Tenant details" },
          "404": { "description": "Tenant not found or not accessible" }
        }
      }
    },
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_createPost",
        "summary": "Create a church-internal post",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePostDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Post created" }
        }
      },
      "get": {
        "operationId": "PostsController_getPosts",
        "summary": "List paginated posts for current tenant (newest first)",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Array of posts with pagination metadata" }
        }
      }
    },
    "/api/posts/global": {
      "post": {
        "operationId": "PostsController_createGlobalPost",
        "summary": "Create a global post (visible to all users, triggers fan-out)",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePostDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Global post created and fan-out queued" }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "summary": "Get a single post by ID",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Post object" },
          "404": { "description": "Post not found or not in current tenant" }
        }
      },
      "patch": {
        "operationId": "PostsController_updatePost",
        "summary": "Update post content (author only)",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePostDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated post" },
          "404": { "description": "Post not found or not the author" }
        }
      },
      "delete": {
        "operationId": "PostsController_deletePost",
        "summary": "Delete a post (author or admin)",
        "tags": ["Posts"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Post deleted" },
          "404": { "description": "Post not found or not authorized" }
        }
      }
    },
...
    "/api/search/posts": {
      "get": {
        "operationId": "SearchController_searchPosts",
        "summary": "Full-text search posts in current tenant (ranked by relevance)",
        "tags": ["Search"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 }, "description": "Full-text search query (supports phrases, OR, exclusion)" },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated search results with cursor" }
        }
      }
    },
    "/api/search/members": {
      "get": {
        "operationId": "SearchController_searchMembers",
        "summary": "Full-text search members in current tenant (by name/email)",
        "tags": ["Search"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 }, "description": "Search by full_name or email" },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated search results with cursor" }
        }
      }
    },
    "/api/stripe/connect/onboard": {
      "post": {
        "operationId": "StripeConnectController_onboard",
        "summary": "Initiate Stripe Connect onboarding (admin only, idempotent)",
        "tags": ["Stripe Connect"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OnboardConnectDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns Stripe AccountLink URL and stripeAccountId" },
          "400": { "description": "Not an admin or no tenant context" }
        }
      }
    },
    "/api/stripe/connect/status": {
      "get": {
        "operationId": "StripeConnectController_getStatus",
        "summary": "Get Stripe Connect onboarding status (admin only)",
        "tags": ["Stripe Connect"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Returns status, chargesEnabled, payoutsEnabled, detailsSubmitted" },
          "400": { "description": "Not an admin or no tenant context" }
        }
      }
    },
...
    "/api/webhooks/stripe": {
      "post": {
        "operationId": "StripeWebhookController_handleStripeWebhook",
        "summary": "Stripe webhook receiver (signature-verified, no JWT)",
        "tags": ["Webhooks"],
        "responses": {
          "200": { "description": "{ received: true }" },
          "401": { "description": "Invalid or missing stripe-signature header" }
        }
      }
    },
...
  },
  "info": {
    "title": "ChurchApp Platform API",
    "description": "Multi-tenant church management platform & global social network.\n\n**Authentication:** All protected endpoints require a Supabase JWT in the `Authorization: Bearer <token>` header.\n\n**Tenant Context:** Most endpoints are tenant-scoped via PostgreSQL Row-Level Security. Call `POST /api/auth/switch-tenant` followed by `POST /api/auth/refresh` to set the active tenant in your JWT.\n\n**Rate Limits:** Global: 100 req/min per IP. Auth endpoints: 5 req/min per IP. Webhooks: unlimited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase JWT access token"
      }
    },
    "schemas": {
...
      "CreateTenantDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 255, "example": "First Baptist Church" }
        }
      },
...
      "CreatePostDto": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string", "maxLength": 5000, "example": "Excited for Sunday service!" },
          "videoMuxPlaybackId": { "type": "string", "description": "Mux playback ID for an attached video" },
          "mentions": { "type": "array", "items": { "type": "string", "format": "uuid" }, "description": "User UUIDs mentioned in the post" }
        }
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "content": { "type": "string", "maxLength": 5000, "example": "Updated post content" }
        }
      },
...
      "OnboardConnectDto": {
        "type": "object",
        "required": ["refreshUrl", "returnUrl"],
        "properties": {
          "refreshUrl": { "type": "string", "format": "uri", "example": "https://app.example.com/onboard", "description": "Redirect URL if onboarding link expires" },
          "returnUrl": { "type": "string", "format": "uri", "example": "https://app.example.com/dashboard", "description": "Redirect URL after onboarding completes" }
        }
      },
...
    }
  }
}